+++ src/components/modules/Question.vue
@@ -1,0 +1,48 @@
+<template>
+  <q-card color="dark">
+    <div v-show="!data.editing || data.editing !== $firebase.auth.currentUser.uid">
+      <q-card-title>
+        <q-icon v-show="!data.editing" class="float-right cursor-pointer" name="fas fa-ellipsis-v" color="primary" size="1rem">
+          <q-popover>
+            <q-list>
+              <q-item link v-close-overlay @click.native="edit(id)">Edit</q-item>
+              <q-item link @click.native="remove(id)">Delete</q-item>
+            </q-list>
+          </q-popover>
+        </q-icon>
+        <span class="float-right" style="font-size: .8rem; vertical-align: top; line-height: 1rem;">{{ data.time }} minutes&nbsp;&nbsp;&nbsp;</span>
+        <q-icon name="fas fa-question" color="primary" size="2rem" />&nbsp;&nbsp;&nbsp;
+        {{ data.text }}
+      </q-card-title>
+    </div>
+    <div v-if="data.editing === $firebase.auth.currentUser.uid">
+      <q-card-main>
+        <q-icon link class="float-right cursor-pointer" name="fas fa-times" color="primary" size="1rem" @click.native="edit('')" />
+        <div class="row gutter-sm">
+          <div class="col-12">
+            <q-input v-model="data.title" float-label="Title" dark />
+          </div>
+          <div class="col-12">
+            <q-input type="number" v-model="data.time" float-label="Estimated Time (in minutes)" dark />
+          </div>
+          <div class="col-12">
+            <q-input v-model="data.text" float-label="Question" type="textarea" :max-height="100" :min-rows="1" dark />
+          </div>
+        </div>
+      </q-card-main>
+    </div>
+  </q-card>
+</template>
+
+<script>
+export default {
+  name: 'mod-text',
+  props: [ 'id', 'data', 'edit', 'remove' ],
+  data () {
+    return {}
+  }
+}
+</script>
+
+<style>
+</style>